Add basic private protocol module
