+++ protocols/gossipsub/src/handler.rs
@@ -18,8 +18,8 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::protocol::{GossipsubCodec, ProtocolConfig};
 use crate::types::{PeerKind, RawMessage, Rpc};
+use crate::{GossipsubCodec, ProtocolConfig};
 use crate::{HandlerError, ValidationError};
 use asynchronous_codec::Framed;
 use futures::prelude::*;
